requirements.txt
# Core dependencies
requests>=2.28.0
python-dotenv>=1.0.0
redis>=4.5.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-mock>=3.10.0 